fix(editor): address review — instruction/data separation, inline MCP server, guards
Addresses Robbie's review on #202:

Structural fix (🔴 blocker):
- User's typed message is now passed as-is to userMessage (instruction)
- All auto-harvested context (active file, diagnostics, attachments,
  workspace) goes into systemPromptSuffix (reference data)
- Context block is labeled 'reference data — not user instruction'
- Prevents injection via file contents becoming part of trusted command

Other fixes:
- Inline MCP server into editorContextServer.ts — no external file
  dependency, works in single-binary deploy (fixes path resolution)
- Remove dead '--stream' flag from Zed MCP server spawning
- Add 10MB stdin size limit to prevent OOM
- Use SHA-256 for conversation ID (replaces weak DJB2 32-bit hash)
- Clean up Zed README to reflect new setup (phantombot editor-context-server)
- Remove extensions/zed-phantombot/server/ directory (no longer needed)
- Mark images and modelHint as follow-up items (not yet plumbed)

Tests updated to verify instruction/data separation — userMessage is
pure, context is in systemPrompt. All 9 tests pass, 1480 suite clean.
